moved the logic from debug:twig to the Twig bridge
